Created separate component for New Post form
New Post is moved into a separate component with a prop currentUser of type Object. The user object is passed from the parent. The new post submitted is then added to currentUser.posts array. Two values are emitted from the child component: this.currentUser object (with new post added) and the new post count Number this.currentUser.posts.lenght. Both are collcted in parent component and displayed.